refactor: change typeorm entities to interfaces
