+++ src/ai/interfaces/chat-message.interface.ts
@@ -1,0 +1,3 @@
+export interface ChatMessage {
+  id: string;
+}
