+++ src/studios/interfaces/studio-membership.interface.ts
@@ -1,5 +1,5 @@
-import { User } from 'src/users/user.interface';
-import { Studio } from 'src/studios/studio.interface';
+import { User } from 'src/users/interfaces/user.interface';
+import { Studio } from 'src/studios/interfaces/studio.interface';
 
 export interface StudioMembership {
   userId: string;
